agentic_backend/agents.py
import os
from textwrap import dedent
from agno.agent import Agent
from agno.team import Team
from agno.models.openai.like import OpenAILike
from agno.tools.duckduckgo import DuckDuckGoTools
from agno.tools.firecrawl import FirecrawlTools
from dotenv import load_dotenv
load_dotenv()
# --- Model Configurations ---

nebius_model = OpenAILike(
    id="openai/gpt-oss-120b",
    api_key=os.getenv("NEBIUS_API_KEY"),
    base_url="https://api.studio.nebius.com/v1/"
)

# --- Custom Tool Placeholders ---

class TwitterTools:
    """Placeholder for Twitter data integration."""
    def get_sentiment(self, token_symbol: str):
        """
        Returns mock sentiment data for a token, including bot detection.
        - Simulates fetching tweets about a token.
        - Flags accounts posting more than 3 times as bots.
        - Returns a summary of authentic sentiment.
        """
        print(f"Fetching Twitter sentiment for {token_symbol}...")
        # Mock data: a list of tuples (account_id, tweet_content)
        mock_tweets = [
            ("user_a", f"So bullish on ${token_symbol}! To the moon! 🚀"),
            ("user_b", f"I'm selling all my ${token_symbol}. This project is a scam."),
            ("bot_1", f"Buy ${token_symbol} now! Guaranteed 100x! #crypto"),
            ("bot_1", f"Don't miss out on ${token_symbol}! #altcoin"),
            ("user_c", f"Just read the whitepaper for ${token_symbol}. Very impressive tech."),
            ("bot_1", f"${token_symbol} is the future of finance! #DeFi"),
            ("bot_1", f"Join the ${token_symbol} revolution!"),
            ("user_d", f"Feeling uncertain about ${token_symbol}'s recent price action."),
        ]

        from collections import Counter
        post_counts = Counter(tweet[0] for tweet in mock_tweets)
        
        authentic_tweets = [
            tweet for tweet in mock_tweets if post_counts[tweet[0]] <= 3
        ]
        
        bullish_count = sum(1 for _, content in authentic_tweets if "bullish" in content or "impressive" in content)
        bearish_count = sum(1 for _, content in authentic_tweets if "scam" in content or "uncertain" in content)

        return {
            "authentic_sentiment": "Bullish" if bullish_count > bearish_count else "Bearish",
            "bullish_posts": bullish_count,
            "bearish_posts": bearish_count,
            "total_authentic_posts": len(authentic_tweets),
            "detected_bots": [user for user, count in post_counts.items() if count > 3]
        }

class GameTheoryTools:
    """Placeholder for Game Theory analysis of market dynamics."""
    def analyze_market_behavior(self, whale_data: dict, sentiment_data: dict):
        """
        Analyzes market behavior from a game theory perspective.
        - Checks for coordination between whales.
        - Models potential outcomes of sentiment vs. whale action.
        """
        print("Applying game theory analysis to market behavior...")
        # This would contain complex logic in a real implementation
        if sentiment_data.get("authentic_sentiment") == "Bullish" and whale_data.get("concentration_risk") < 0.2:
            return "Positive-sum game: Conditions are favorable for retail and whales."
        else:
            return "Zero-sum game: High risk of manipulation or conflict of interest."

class PythNetworkTools:
    """Placeholder for Pyth Network integration to fetch live token prices."""
    def get_price(self, token_id: str):
        """Fetches the real-time price of a token."""
        print(f"Fetching price for {token_id}...")
        pass

class TheGraphTools:
    """Placeholder for The Graph integration to query substreams for whale data."""
    def query_whale_activity(self, contract_address: str):
        """Queries on-chain activity for a given contract address."""
        print(f"Querying whale activity for {contract_address}...")
        pass

class PyEthTools:
    """Placeholder for py-eth integration for on-chain data."""
    def get_market_cap(self, token_id: str):
        """Fetches the market cap of a token."""
        print(f"Fetching market cap for {token_id}...")
        pass

# --- Agent Definitions ---

discovery_agent = Agent(
    name="Token Discovery & Intelligence Agent",
    model=nebius_model,
    tools=[FirecrawlTools(api_key=os.getenv("FIRECRAWL_API_KEY")), TwitterTools()],
    description="An elite crypto market researcher specializing in token discovery, funding round analysis, and early-stage project evaluation.",
    instructions=dedent("""
        - Focus on web scraping and crawling to discover newly funded utility tokens from platforms like CryptoRank.
        - Extract comprehensive tokenomics, investor profiles, and vesting schedules from project websites.
        - Assess the token's utility and competitive landscape based on available documentation.
        - Use Twitter sentiment analysis to gauge authentic community bullishness while filtering out bot activity.
        - Provide comprehensive analysis based on the data you can successfully gather from web sources.
    """),
    expected_output="A comprehensive report on newly discovered tokens with their investment potential and social sentiment.",
)

whale_agent = Agent(
    name="Whale Tracking & Behavior Analysis Agent",
    model=nebius_model,
    tools=[TheGraphTools()],
    description="A specialized on-chain analyst with expertise in whale wallet identification and large transaction monitoring.",
    instructions=dedent("""
        - Identify the top 100 holder wallets for a given token.
        - Monitor real-time large transfers exceeding predefined thresholds.
        - Analyze whale accumulation and distribution patterns.
        - Assess concentration risk for potential market manipulation.
    """),
    expected_output="An intelligence report on whale activity, concentration risk, and potential market movements.",
)

market_agent = Agent(
    name="Live Market Data & Technical Analysis Agent",
    model=nebius_model,
    tools=[PythNetworkTools(), PyEthTools()],
    description="A quantitative market analyst specializing in real-time price tracking and market capitalization analysis.",
    instructions=dedent("""
        - Deliver live token prices and market capitalization.
        - Analyze trading volume and liquidity depth.
        - Provide technical indicators like RSI and MACD.
        - Assess market sentiment from price action.
    """),
    expected_output="A real-time market intelligence dashboard with key metrics and technical analysis.",
)

financial_agent = Agent(
    name="Financial Analysis & Investment Decision Agent",
    model=nebius_model,
    tools=[GameTheoryTools()],
    description="A senior crypto financial analyst with expertise in investment thesis development and risk modeling.",
    instructions=dedent("""
        - Synthesize all collected data into a professional investment recommendation.
        - Analyze fundamental valuation and risk-adjusted return projections.
        - Apply game theory principles to assess the strategic landscape of market participants.
        - Provide portfolio allocation guidance and scenario analysis.
        - Formulate a clear investment thesis with risk management strategies.
    """),
    expected_output="A professional investment report with a clear BUY/SELL/HOLD recommendation, detailed financial analysis, and strategic market insights.",
)

# --- Team Definition ---

finrizz_team = Team(
    name="FinRizz Analysis Team",
    members=[discovery_agent, whale_agent, market_agent, financial_agent],
    instructions=[
        "You are the FinRizz financial analysis team coordinator.",
        "Process all agent inputs sequentially for comprehensive token analysis.",
        "Synthesize the findings into a clear, actionable investment recommendation."
    ],
    show_members_responses=True,
    markdown=True
)
