agentic_backend/requirements.txt
python-telegram-bot
python-dotenv
agno
openai
google-generativeai
firecrawl-py
lxml_html_clean
ddgs

# Core requirements
# requests>=2.28.0

# Optional requirements for advanced features
# python-dotenv>=1.0.0
# pandas>=1.5.0
# aiohttp>=3.8.0
# asyncio-throttle>=1.0.0

# # Web API requirements
# flask>=2.2.0
# flask-cors>=4.0.0

# # Development requirements
# pytest>=7.0.0
# pytest-asyncio>=0.21.0

# # telegram bot & dotenv
# python-telegram-bot==20.7
# python-dotenv==1.0.1


